Move test_join_order.py to test_merge.py
